requirements.txt
python-telegram-bot==20.0
requests==2.31.0
python-dotenv==1.0.1
supabase-py
cloudinary==1.36.0
httpx==0.24.1